refactor pipeline code
