--- colossalai/inference/cai_engine/__init__.py
@@ -1,3 +1,0 @@
-from .engine import CaiInferEngine
-
-__all__ = ['CaiInferEngine']
