+++ colossalai/inference/pipeline/batch_infer_state.py
