+++ colossalai/inference/pipeline/modeling/llama.py
