+++ colossalai/inference/pipeline/policies/__init__.py
